adding in tinydb, splitting out app files a bit more
