+++ .gitignore
@@ -1,5 +1,5 @@
 /data/output
-
+/db.json
 .env
 
 logs
